+++ Samples/Sample-SSHWindowsService.ps1
@@ -3,7 +3,6 @@
 $modulePath = "D:\Projects\PSNovusTools\PSNovusTools"
 Import-Module -Name $modulePath -Force
 
-
 $remoteServer = "10.0.0.21"
 $remoteUser = "Administrator"
 $remotePassword = Convert-StringToSecureString("w0rk1ngh@rd")
@@ -11,3 +10,7 @@ $remoteServiceName = "docker"
 
 
 Get-SSHWindowsService -remoteServer $remoteServer -remoteUser $remoteUser -remotePassword $remotePassword -remoteServiceName $remoteServiceName
+
+Start-SSHWindowsService -remoteServer $remoteServer -remoteUser $remoteUser -remotePassword $remotePassword -remoteServiceName $remoteServiceName
+
+Stop-SSHWindowsService -remoteServer $remoteServer -remoteUser $remoteUser -remotePassword $remotePassword -remoteServiceName $remoteServiceName